+++ xllm/core/util/env_var.cpp
@@ -34,24 +34,24 @@ bool get_bool_env(const std::string& key, bool defaultValue) {
           strVal == "True");
 }
 
-int get_int_env(const std::string& key, int defaultValue) {
+int64_t get_int_env(const std::string& key, int64_t defaultValue) {
   const char* val = std::getenv(key.c_str());
   if (val == nullptr) {
     return defaultValue;
   }
   // Use strtol for proper error handling
   char* endptr;
-  long int result = std::strtol(val, &endptr, 10);
+  int64_t result = std::strtol(val, &endptr, 10);
   // Check if conversion was successful (endptr points to end of string or valid
   // terminator)
   if (endptr == val || *endptr != '\0') {
     return defaultValue;
   }
   // Check for overflow/underflow
-  if (result < INT_MIN || result > INT_MAX) {
+  if (result < INT64_MIN || result > INT64_MAX) {
     return defaultValue;
   }
-  return static_cast<int>(result);
+  return result;
 }
 
 std::string get_string_env(const std::string& name) {
@@ -62,11 +62,11 @@ std::string get_string_env(const std::string& name) {
   return std::string(val);
 }
 
-int get_process_group_test_timeout_seconds() {
+int64_t get_process_group_test_timeout_seconds() {
   // Default timeout is 4 seconds, but can be overridden via environment
   // variable to accommodate multi-node multi-device communication scenarios
   // where network latency may require a longer timeout period.
-  constexpr int kDefaultTimeoutSeconds = 4;
+  constexpr int64_t kDefaultTimeoutSeconds = 4;
   constexpr const char* kTimeoutEnvVar =
       "XLLM_PROCESS_GROUP_ASYNC_TIMEOUT_SECONDS";
   return get_int_env(kTimeoutEnvVar, kDefaultTimeoutSeconds);
